Looping Statement
A for loop is used to iterate over a sequence like lists, type, dictionaries, sets, or even strings. Loop statements will be executed for each item of the sequence. Takes the first item of the iterable, executes the statement, and moves the pointer to the next item until it reaches the last item of the sequence.
In computer science, a for-loop (or simply for loop) is a control flow statement for specifying iteration, which allows code to be executed repeatedly. Various keywords are used to specify this statement: descendants of ALGOL use "for", while descendants of Fortran use "do".